:art: Update web3-eth packageto export utility functions
